frontend: Fix include cycles in Tables.
